Fix issue with db

+++ src/db.ts
@@ -4,7 +4,7 @@ import { Config } from 'node-json-db/dist/lib/JsonDBConfig';
 
 export const db = new JsonDB(new Config('db', true, true, '/'));
 
-const OPEN_ORDERS_PATH = '/open_orders';
+export const OPEN_ORDERS_PATH = '/open_orders';
 
 export const addOpenOrder = (
   symbol: string,
@@ -18,14 +18,18 @@ export const addOpenOrder = (
 };
 
 export const getOpenOrders = (symbol: string) => {
-  return db.getData(`${OPEN_ORDERS_PATH}/${symbol}`) as OpenOrder[];
+  if (db.exists(`${OPEN_ORDERS_PATH}/${symbol}`))
+    return db.getData(`${OPEN_ORDERS_PATH}/${symbol}`) as OpenOrder[];
 };
 
 export const deleteOpenOrder = (symbol: string, orderId: number) => {
-  const index = db.getIndex(`${OPEN_ORDERS_PATH}/${symbol}`, orderId, 'id');
-  db.delete(`${OPEN_ORDERS_PATH}/${symbol}[${index}]`);
+  if (db.exists(`${OPEN_ORDERS_PATH}/${symbol}`)) {
+    const index = db.getIndex(`${OPEN_ORDERS_PATH}/${symbol}`, orderId, 'id');
+    db.delete(`${OPEN_ORDERS_PATH}/${symbol}[${index}]`);
+  }
 };
 
 export const deleteOpenOrders = (symbol: string) => {
-  db.delete(`${OPEN_ORDERS_PATH}/${symbol}`);
+  if (db.exists(`${OPEN_ORDERS_PATH}/${symbol}`))
+    db.delete(`${OPEN_ORDERS_PATH}/${symbol}`);
 };
